Update README.md
The aim of the “Roadwatch” is to create an effective system for automatically identifying and categorizing 
traffic signs from images or video feeds. By harnessing machine learning methodologies, the project endeavors 
to bolster road safety by promptly recognizing traffic signs, thereby aiding drivers in adhering to traffic 
regulations and reducing the likelihood of accidents due to human error or oversight. To address the variability 
of road signs across diverse regions and environments, the project will train models on varied datasets to 
ensure consistent performance in different conditions. Furthermore, the project seeks to assess the practical 
application of the system in scenarios such as autonomous vehicles, advanced driver-assistance systems 
(ADAS), or traffic management platforms to contribute to the advancement of intelligent   transportation 
systems.